Add files via upload
Changed "FileUtils" in make.rb to "fileutils" as Linux is case sensitive and required the include to be all lower case.

Modified README.md to point towards the acme compiler on github, which compiles and works under Linux (Ubuntu 16.04 tested).

Modified README.md to state that the program should be called with either "ruby make.rb game.z5" or the make.sh script which simply does the aforementioned.

Changed make.rb so that it accepts z[1-8] as an argument. Also modified it so that when .dat files are passed with a -z[1-8] the passed parameter kicks in and the d64 file gets made, instead of failing.

I hacked and slashed at some stuff here. However, it seems to work under Linux mostly now.