Fix reverse video in buffered output
Without this the "styled text" test in etude.z5 doesn't correctly show
reverse video. The problem is that the reverse video and normal video
control codes get processed by s_printchar straight away, while the
intervening text is buffered.

This commit adds a parallel print_buffer2 which records the status of
s_reverse for every character buffered, then restores that status
temporarily when the buffered output is passed to s_printchar. Doing it
this way avoids the complication of having zero-width control characters
in print_buffer.